change login process
