+++ main.py
@@ -4,11 +4,9 @@
 import streamlit as st
 import pandas as pd
 import altair as alt
-# from streamlit_local_storage import LocalStorage
-from streamlit_local_storage import LocalStorage
 
-
-from request import is_authenticate, vasahm_query, get_nonce, get_key
+from login import check_local_token, login
+from request import vasahm_query
 from menu import add_menu
 
 st.set_page_config(layout='wide',
@@ -16,7 +14,6 @@ st.set_page_config(layout='wide',
                     page_icon="./assets/favicon.ico",
                     initial_sidebar_state='expanded')
 st.session_state.ver = '0.1.5'
-sessionBrowserS = LocalStorage()
 
 with open( "style.css", encoding="utf-8") as css:
     st.markdown( f'<style>{css.read()}</style>' , unsafe_allow_html= True)
@@ -44,7 +41,7 @@ HTML = """<!DOCTYPE html>
 
 <body>
   <div id="content">
-    <p>وبسایت در مرحله آزمایشی است، از اطلاعات با احتیاط استفاده کنید. در صورت مشاهده ایراد میتوانید به <a href="https://t.me/Khiaboon_Hafez_Admin" target="_blank">@Khiaboon_Hafez_Admin</a> پیام دهید.</p>
+    <p>وبسایت در مرحله آزمایشی است، از اطلاعات با احتیاط استفاده کنید. در صورت مشاهده ایراد میتوانید به تست<a href="https://t.me/Khiaboon_Hafez_Admin" target="_blank">@Khiaboon_Hafez_Admin</a> پیام دهید.</p>
   </div>
 </body>
 
@@ -54,46 +51,9 @@ st.components.v1.html(HTML, height=60, scrolling=False)
 # st.sidebar.image(image="./assets/logo.png")
 st.sidebar.header(f'Vasahm DashBoard `{st.session_state.ver}`')
 
-
-def get_email_callback():
-    """Send nonce to entered email."""
-    has_error, message = get_nonce(st.session_state.email)
-    if has_error:
-        st.error(message, icon="🚨")
-    else:
-        submit_nonce = st.form("submit_nonce")
-        submit_nonce.text_input('کد تایید خود را وارد کنید',
-                                placeholder='XXXX',
-                                key="nonce")
-        submit_nonce.form_submit_button("ارسال", on_click = get_nonce_callback )
-
-def get_nonce_callback():
-    """Confirm nonce for login."""
-    has_error, message = get_key(st.session_state.email, st.session_state.nonce)
-    if has_error:
-        st.error(message, icon="🚨")
-        del st.session_state["nonce"]
-    else:
-        st.session_state["token"] = message
-        sessionBrowserS.setItem("saved_token", message)
-
-sessionBrowserS.getItem("saved_token", key='temp1')
-if st.session_state.temp1 is not None:
-    if "storage" in st.session_state.temp1:
-        if st.session_state.temp1['storage'] is not None:
-            saved_token = st.session_state.temp1['storage']['value']
-            if is_authenticate(saved_token):
-                st.session_state["token"] = saved_token
-            else:
-                sessionBrowserS.deleteItem("saved_token")
-
+check_local_token()
 if "token" not in st.session_state:
-    get_email = st.form("get_email")
-    email = get_email.text_input('ایمیل خود را وارد کنید',
-                                 placeholder='example@mail.com',
-                                 key="email")
-    # Every form must have a submit button.
-    submitted = get_email.form_submit_button("دریافت کد", on_click = get_email_callback )
+    login()
 else:
     df = pd.read_csv("data.csv").dropna()
     list_of_name = df['name'].to_list()
@@ -109,7 +69,7 @@ else:
             order by 
                 "stockData".id desc
             """
-    
+
     error, stock_data = vasahm_query(query_string)
     if error:
         st.error(stock_data, icon="🚨")
